+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Add small example to `README.md` and crate root documentation
  - Add `uint8`/`int8` support to the bitround codec
  - Add `len()` and `is_empty()` methods to array subset iterator producers
+ - Add `Array::chunk_origin()`
 
 ### Changed
  - **Breaking**: Bump MSRV to 1.73 (5 October, 2023)